+++ app/src/main/java/com/groundsoft/dean/shoppinglist/Models/DbConsts.java
@@ -1,7 +1,6 @@
 package com.groundsoft.dean.shoppinglist.Models;
 
 import android.database.sqlite.SQLiteDatabase;
-import com.groundsoft.dean.shoppinglist.SQLite;
 
 class DbConsts {
     static final int DATABASE_VERSION = 1;
@@ -11,10 +10,10 @@ class DbConsts {
 
         //CREATE TABLE IF NOT EXISTS
 
-        String query = "Create table " + SQLite.TABLE_LISTS + " (" +
-                SQLite.TABLE_LISTS_KEY_ID + " integer primary key, " +
-                SQLite.TABLE_LISTS_KEY_NAME + " text, " +
-                SQLite.TABLE_LISTS_KEY_DATE + " integer)";
+        String query = "Create table " + Lists.TABLE_LISTS + " (" +
+                Lists.TABLE_LISTS_KEY_ID + " integer primary key, " +
+                Lists.TABLE_LISTS_KEY_NAME + " text, " +
+                Lists.TABLE_LISTS_KEY_DATE + " integer)";
         db.execSQL(query);
 
         String query2 = "Create table " + Items.TABLE_ITEMS + " (" +
@@ -34,10 +33,10 @@ class DbConsts {
                 Categories.TABLE_CATEGORIES_KEY_ORDER + " integer)";
         db.execSQL(query3);
 
-        String query4 = "Create table " + SQLite.TABLE_DEFITEMS + " (" +
-                SQLite.TABLE_DEFITEMS_KEY_ID + " integer primary key, " +
-                SQLite.TABLE_DEFITEMS_KEY_NAME + " text, " +
-                SQLite.TABLE_DEFITEMS_KEY_ORDER + " integer)";
+        String query4 = "Create table " + DefItems.TABLE_DEFITEMS + " (" +
+                DefItems.TABLE_DEFITEMS_KEY_ID + " integer primary key, " +
+                DefItems.TABLE_DEFITEMS_KEY_NAME + " text, " +
+                DefItems.TABLE_DEFITEMS_KEY_ORDER + " integer)";
         db.execSQL(query4);
 
     }
@@ -45,8 +44,8 @@ class DbConsts {
     static void upgradeAll(SQLiteDatabase db, int oldVersion, int newVersion) {
         db.execSQL("drop table if exists " + Items.TABLE_ITEMS);
         db.execSQL("drop table if exists " + Categories.TABLE_CATEGORIES);
-        db.execSQL("drop table if exists " + SQLite.TABLE_LISTS);
-        db.execSQL("drop table if exists " + SQLite.TABLE_DEFITEMS);
+        db.execSQL("drop table if exists " + Lists.TABLE_LISTS);
+        db.execSQL("drop table if exists " + DefItems.TABLE_DEFITEMS);
         createAll(db);
     }
 }
